pyproject.toml
[project]
name = "flask-structure"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Jhaendar", email = "lycaon.anderson@gmail.com" }]
dependencies = ["flask>=3.0.2", "flask-sqlalchemy>=3.1.1", "requests>=2.31.0"]
readme = "README.md"
requires-python = ">= 3.8"

# [project.scripts]
# hello = "flask_structure:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pre-commit>=3.5.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "djlint>=1.34.1",
]
[tool.rye.scripts]
# These three options are equivalent:
flask_run = { cmd = "flask run --debug", env = { FLASK_APP = "flask_structure.app", FLASK_ENV = "development", SECRET_KEY = "dev" } }

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/flask_structure"]
